requirements.txt
python-telegram-bot==21.3
openai>=1.0.0
python-pptx>=0.6.21
